adds cart functionality
