+++ templates/demo-store-neue/index.html
@@ -9,6 +9,8 @@
     <title>Hydrogen Demo Store</title>
     <link rel="stylesheet" href="/src/styles/index.css" />
     <link rel="preconnect" href="https://cdn.shopify.com" />
+    <link rel="preconnect" href="https://shop.app/" />
+    <link rel="preconnect" href="https://hydrogen-preview.myshopify.com/" />
   </head>
   <body>
     <div id="root"></div>
